Struct

+++ VI/VI.Genetic/Fitness/IFitnessFunction.cs
@@ -4,6 +4,6 @@ namespace VI.Genetic.Fitness
 {
     public interface IFitnessFunction
     {
-        double Evaluate(IChromosome chromosome);
+        float Evaluate(IChromosome chromosome);
     }
 }
